Update modules
fix free error

the issue was we were performing a double free, one process was running
and ask to be free'd. Another showed up and was worst than the current
process so the current process was asked to run again causing a double
free

this is fixed now as i have changed that when we sort we sort on ahead
of buf_tail, this ensures we will not consider the currently running
process in the sorting algorithm

however, if we have an arrival rate of zero do not perform this +1 sort.
if we were to do this we would not properly sort the batch job.
For instance if all 5 jobs arrive at the same time and job 2 is better than
job 1. If we perform a sort of buf_tail + 1 we will sort all the jobs except
the first job, even though it is worst than the second job.